adding watchlist fetch
The access token is needed to get watchlist from auth service, I tried in previous commits the token generation, but for some reason, we are unable to get the access token, as new type WalletClient inside web3auth service is not finding the getAdresses method.

Still looking a for a way to make the fetch without needing the access token or getting it.